NEO4J/etl/requirements.txt
neo4j>=5.14.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
